requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
botocore==1.20.109
chardet==4.0.0
discord.py==1.7.3
idna==3.2
jmespath==0.10.0
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
